speedometer changed to chart js and ui implemented on homepage
