+++ TranSafe/package.json
@@ -31,6 +31,7 @@
     "@ionic-native/splash-screen": "^5.0.0",
     "@ionic-native/status-bar": "^5.0.0",
     "@ionic/angular": "^5.0.0",
+    "angular-gauge-chart": "^0.7.2",
     "chart.js": "^2.9.3",
     "child_process": "^1.0.2",
     "core-js": "^2.5.4",
