+++ TranSafe/src/app/home/home.module.ts
@@ -1,4 +1,4 @@
-import { SpeedometerComponent } from './speedometer/speedometer.component';
+
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
@@ -6,7 +6,7 @@ import { FormsModule } from '@angular/forms';
 import { HomePage } from './home.page';
 
 import { HomePageRoutingModule } from './home-routing.module';
-
+import { GaugeChartModule } from 'angular-gauge-chart';
 
 
 @NgModule({
@@ -14,8 +14,9 @@ import { HomePageRoutingModule } from './home-routing.module';
     CommonModule,
     FormsModule,
     IonicModule,
+    GaugeChartModule,
     HomePageRoutingModule,
   ],
-  declarations: [HomePage, SpeedometerComponent]
+  declarations: [HomePage]
 })
 export class HomePageModule {}
