+++ TranSafe/src/app/home/home.page.scss
@@ -16,3 +16,38 @@ ion-card-content {
 ion-menu-button {
   color: rgb(29, 29, 29);
 }
+
+
+.my-css-class {
+  background-color: rgb(65, 65, 65)!important;;
+  --border-radius: 0;
+  --box-shadow: 0;
+
+}
+
+
+.track-button {
+  position: absolute;
+  left: 78%;
+  bottom: 2%;
+  width: 60px;
+  height: 30px;
+}
+
+.speedometer{
+  width: fit-content;
+  height: fit-content;
+  left: 0;
+  right: 0;
+  margin: auto;
+}
+
+
+.speed-value {
+  font-size: 1em;
+  color: rgb(36, 36, 36);
+  top: 90px;
+  left: 35%;
+  right: 0;
+  position: absolute;
+}
