--- TranSafe/src/app/home/speedometer/speedometer.component.html
@@ -1,22 +1,0 @@
-<div class="speedometer-card">
-  <div class="speedometer">
-    <div class="num num1"><div>110</div></div>
-    <div class="num num2"><div>130</div></div>
-    <div class="num num3"><div>150</div></div>
-    <div class="num num4"><div>170</div></div>
-    <div class="num num5"><div></div></div>
-    <div class="num num6"><div></div></div>
-    <div class="num num7"><div></div></div>
-    <div class="num num8"><div>10</div></div>
-    <div class="num num9"><div>30</div></div>
-    <div class="num num10"><div>50</div></div>
-    <div class="num num11"><div>70</div></div>
-    <div class="num num12"><div>90</div></div>
-
-
-    <div class="current-speed" #currentSpeed ></div>
-    <div class="speed-value"><span class="small">Speed Limit:</span>{{speedlmt}}</div>
-    <div class="unit">km/h</div>
-  </div>
-  <ion-button color="primary" [ngClass]="{'my-css-class': updateBtnText === 'Stop'}"  class="track-button" (click)="startTrackUser()">{{updateBtnText}}</ion-button>
-</div>
