--- TranSafe/src/app/home/speedometer/speedometer.component.scss
@@ -1,190 +1,0 @@
-.speedometer-card {
-
-  width: auto;
-  height: 35vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 5px;
-  overflow: hidden;
-}
-
-
-.speedometer {
-  position: relative;
-  width: 220px;
-  height: 220px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: white;
-  border-radius: 50%;
-  border: 15px solid #5F4591;
-
-}
-
-.speedometer::before {
-  content: '';
-  position: absolute;
-  background-color: rgb(48, 48, 48);
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-}
-
-
-.unit{
-  color: #5F4591;
-
-}
-span.small {
-  font-size: xx-small;
-}
-.speed-value {
-  margin-top: 48px;
-  font-size: 1.5em;
-  color: red;
-}
-
-.track-button {
-  position: absolute;
-  left: 78%;
-  bottom: 2%;
-  width: 60px;
-  height: 30px;
-}
-
-.current-speed {
-  width: 5px;
-  height: 35%;
-  background: rgb(48, 48, 48);
-  transform-origin: 50% 100%;
-  border-radius: 100% 100% 0 0;
-  position: absolute;
-  top: 16%;
-  left: 49.5%;
-  opacity: 0.5;
-
-}
-
-.num {
-  height: 100%;
-  position: absolute;
-  left: calc(50%-0.5em);
-}
-
-.num div {
-  line-height: 2em;
-  color: rgb(133, 131, 131);
-  text-align: center;
-  vertical-align: middle;
-}
-
-// Every 30 degree is equivalent to 20km/s
-
-.num1 {
-  transform: rotate(30deg);
-}
-
-.num1 div {
-  transform: rotate(-30deg);
-}
-
-.num2 {
-  transform: rotate(60deg);
-}
-
-.num2 div {
-  transform: rotate(-60deg);
-}
-
-.num3 {
-  transform: rotate(90deg);
-}
-
-.num3 div {
-  transform: rotate(-90deg);
-}
-
-.num4 {
-  transform: rotate(120deg);
-}
-
-.num4 div {
-  transform: rotate(-120deg);
-}
-
-.num5 {
-  transform: rotate(150deg);
-}
-
-.num5 div {
-  transform: rotate(-150deg);
-}
-
-.num6 {
-  transform: rotate(180deg);
-}
-
-.num6 div {
-  transform: rotate(-180deg);
-}
-
-.num7 {
-  transform: rotate(210deg);
-}
-
-.num7 div {
-  transform: rotate(-210deg);
-}
-
-.num8 {
-  transform: rotate(240deg);
-}
-
-.num8 div {
-  transform: rotate(-240deg);
-}
-
-.num9 {
-  transform: rotate(270deg);
-}
-
-.num9 div {
-  transform: rotate(-270deg);
-}
-
-.num10 {
-  transform: rotate(300deg);
-}
-
-.num10 div {
-  transform: rotate(-300deg);
-}
-
-.num11 {
-  transform: rotate(330deg);
-}
-
-.num11 div {
-  transform: rotate(-330deg);
-}
-
-.num12 {
-  transform: rotate(360deg);
-}
-
-.num12 div {
-  transform: rotate(-360deg);
-}
-
-.my-css-class {
-  background-color: rgb(65, 65, 65)!important;;
-  --border-radius: 0;
-  --box-shadow: 0;
-
-
-}
-
-
-
